Log: finish pimpl refactor. Refs #33.
Also:
* Style refactor
* Remove unused code
* Document remaining code / add important notes
